[SP-3631] move ConsentDelegateMock to its own file
